Squashed commit of the following:

commit cc15aa3048a4006dcede48ae2c74292f1185ef44
Author: Michael Altizer <mialtize@cisco.com>
Date:   Tue Mar 2 13:03:42 2021 -0500

    log: Add printf format attribute to TextLog_Print() and clean up the fallout

+++ src/codecs/ip/cd_frag.cc
@@ -137,7 +137,7 @@ void Ipv6FragCodec::log(TextLog* const text_log, const uint8_t* raw_pkt,
     const uint16_t offlg = fragh->off();
 
     TextLog_Print(text_log, "Next:0x%02X Off:%u ID:%u",
-        fragh->ip6f_nxt, offlg, fragh->id());
+        static_cast<uint8_t>(fragh->ip6f_nxt), offlg, fragh->id());
 
     if (fragh->mf())
         TextLog_Puts(text_log, " MF");
